Fix build & cleanup
