+++ next/pages/api/eforms/[id]/transform/xmlToJson.ts
@@ -1,5 +1,5 @@
 import { withSentry } from '@sentry/nextjs'
-import { removeNeedlessXmlTransformArraysRecursive, validateFormName } from '@utils/forms'
+import { removeNeedlessXmlTransformArraysRecursive } from '@backend/utils/forms'
 import { forceString } from '@utils/utils'
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { readFile } from 'node:fs/promises'
@@ -16,7 +16,8 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     return res.status(400).json({ message: 'Invalid method or missing "data" field on body' })
   let jsonSchema
   try {
-    const formName = await validateFormName(req.query.id)
+    // TODO change how to look up for the schema
+    const formName = req.query.id
     if (!formName) throw new Error(`Invalid form name - validateFormName returned: ${formName}`)
     const filePath = resolve(cwd(), 'forms', forceString(req.query.id), 'schema.json')
     jsonSchema = JSON.parse(await readFile(filePath, 'utf8'))
